move Kpoint.short() to utilities
